Removed utility.ts in favour of utils.ts, moved contents over and switched imports to match change
