+++ src/lib/utils.ts
@@ -4,3 +4,19 @@ import { twMerge } from 'tailwind-merge';
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+
+export const getDate = () => {
+  const date = new Date();
+  const offsetMs = date.getTimezoneOffset() * 60000; // Convert offset to milliseconds
+  return new Date(date.getTime() - offsetMs);
+};
+
+export const undoConversion = (date: Date) => {
+  const offsetMs = date.getTimezoneOffset() * 60000; // Convert offset to milliseconds
+  return new Date(date.getTime() + offsetMs);
+};
+
+export function getLocalISOString() {
+  return getDate().toISOString().split('T')[0]; //only get the month-day-year
+}
+
